Move JsonObjectTest to its own package
