+++ JavaModules/idispatch-client/src/test/java/net/pkhapps/idispatch/client/v3/util/ColorTest.java
@@ -1,5 +1,6 @@
 package net.pkhapps.idispatch.client.v3.util;
 
+import net.pkhapps.idispatch.client.v3.test.JsonObjectTest;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
